requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
opencv-python
pillow
tensorflow
kagglehub
torch
pytest
qiskit
qiskit-machine-learning
qiskit-ibm-runtime
qiskit_aer
gunicorn
monai
flask
pylatexenc
python-dotenv
